Add encryption roundtrip tests
