+++ rust-enc-ui/Cargo.toml
@@ -13,3 +13,6 @@ getrandom = "0.3.3"
 rfd = "0.15.4"
 thiserror = "2.0.14"
 zeroize = "1.8.1"
+
+[dev-dependencies]
+tempfile = "3.10"
